requirements.txt
aiohttp==3.10.11
APScheduler==3.10.4
jinja2==3.1.4
motor==3.6.0
PyYAML==6.0.2
python-dotenv==1.0.1
python-telegram-bot==21.6
discord-webhook==1.3.1
uvicorn==0.30.6
feedparser==6.0.11
pytest==8.3.3
pytest-asyncio==0.24.0
